fix cart

+++ Ascension/Controllers/AuthenticationController.fs
@@ -84,7 +84,9 @@ type AuthenticationController() =
             then
                 let cartId = this.HttpContext.Session.GetInt32("cartId") |> int
                 let oldCart = context.Cart.FirstOrDefault(fun c -> c.AuthorizedUserId = dbUser.Id)
-                context.Cart.Remove(oldCart) |> ignore
+                if oldCart <> null
+                then
+                    context.Cart.Remove(oldCart) |> ignore
                 context.Cart.First(fun c -> c.Id = cartId).AuthorizedUserId <- dbUser.Id
         context.SaveChanges() |> ignore
